# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WMAX API
  description: Remote Patient Monitoring — production-grade clinical signal backend
  version: 2.0.0
servers:
- url: /
  description: Standart origin (Lokal dev :8000 yoki reverse proxy ildizi)
paths:
...
  /api/v1/patients/{id}:
    get:
      tags:
      - patients
      summary: Patient detail with multi-param series, baselines, trend, and prognosis.
      operationId: get_patient_detail_api_v1_patients__id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
          description: History window in days
          default: 7
          title: Days
        description: History window in days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - patients
      summary: Patch patient profile
      operationId: patch_patient_api_v1_patients__id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Patch Patient Api V1 Patients  Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/patients/{id}/twin/prognosis:
    get:
      tags:
      - ai-clinical
      summary: Get cached or generate AI digital twin clinical prognosis
      description: Evaluates 72-hour clinical prognosis over complete digital twin
        context (Section 8.3 & 8.4).
      operationId: get_patient_twin_prognosis_api_v1_patients__id__twin_prognosis_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinPrognosisInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - ai-clinical
      summary: Generate comprehensive AI digital twin clinical prognosis
      description: Evaluates 72-hour clinical prognosis over complete digital twin
        context (Section 8.3 & 8.4).
      operationId: get_patient_twin_prognosis_api_v1_patients__id__twin_prognosis_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinPrognosisInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/nurse-handover:
    get:
      tags:
      - ai-clinical
      summary: Get cached or generate SBAR nurse handover and shift patrol checklist
      description: Generates an actionable SBAR report and priority checklist for
        on-duty nurses.
      operationId: get_nurse_handover_api_v1_patients__id__nurse_handover_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NurseHandoverSBAR'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - ai-clinical
      summary: Generate AI-powered SBAR nurse handover and shift patrol checklist
      description: Generates an actionable SBAR report and priority checklist for
        on-duty nurses.
      operationId: get_nurse_handover_api_v1_patients__id__nurse_handover_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NurseHandoverSBAR'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Alert:
      properties:
        id:
          type: integer
          title: Id
        ts:
          type: string
          format: date-time
          title: Ts
        level:
          type: string
          enum:
          - green
          - amber
          - red
          - no_data
          title: Level
        composite_score:
          type: number
          title: Composite Score
        triggered_params:
          additionalProperties:
            type: number
          type: object
          title: Triggered Params
          default: {}
        anomaly_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Anomaly Score
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - id
      - ts
      - level
      - composite_score
      title: Alert
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NurseChecklistItem:
      properties:
        id:
          type: string
          title: Id
        task:
          type: string
          title: Task
        priority:
          type: string
          enum:
          - low
          - medium
          - high
          - critical
          title: Priority
        category:
          type: string
          enum:
          - vitals
          - medication
          - device
          - observation
          title: Category
        completed:
          type: boolean
          title: Completed
          default: false
      type: object
      required:
      - id
      - task
      - priority
      - category
      title: NurseChecklistItem
    NurseHandoverSBAR:
      properties:
        situation:
          type: string
          title: Situation
        background:
          type: string
          title: Background
        assessment:
          type: string
          title: Assessment
        recommendation:
          type: string
          title: Recommendation
        shift_checklist:
          items:
            $ref: '#/components/schemas/NurseChecklistItem'
          type: array
          title: Shift Checklist
          default: []
        clinical_urgency:
          type: string
          enum:
          - routine
          - urgent
          - critical
          title: Clinical Urgency
        vital_flags:
          items:
            type: string
          type: array
          title: Vital Flags
          default: []
        confidence_score:
          type: number
          title: Confidence Score
          default: 0.95
        evidence_citations:
          items:
            type: string
          type: array
          title: Evidence Citations
          default: []
        generated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Generated At
      type: object
      required:
      - situation
      - background
      - assessment
      - recommendation
      - clinical_urgency
      title: NurseHandoverSBAR
    ParamSeries:
      properties:
        param:
          type: string
          title: Param
        points:
          items:
            $ref: '#/components/schemas/SeriesPoint'
          type: array
          title: Points
        baseline_median:
          anyOf:
          - type: number
          - type: 'null'
          title: Baseline Median
        baseline_low:
          anyOf:
          - type: number
          - type: 'null'
          title: Baseline Low
        baseline_high:
          anyOf:
          - type: number
          - type: 'null'
          title: Baseline High
        deviated_ranges:
          items:
            additionalProperties:
              type: string
              format: date-time
            type: object
          type: array
          title: Deviated Ranges
          default: []
      type: object
      required:
      - param
      - points
      title: ParamSeries
...
    PatientDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        full_name:
          type: string
          title: Full Name
        age:
          type: integer
          title: Age
        sex:
          anyOf:
          - type: string
            enum:
            - m
            - f
          - type: 'null'
          title: Sex
        diagnosis:
          type: string
          title: Diagnosis
        district:
          type: string
          title: District
        phase:
          type: string
          enum:
          - calib
          - learning
          - full
          title: Phase
        level:
          type: string
          enum:
          - green
          - amber
          - red
          - no_data
          title: Level
        composite_score:
          type: number
          title: Composite Score
        triggered_params:
          additionalProperties:
            type: number
          type: object
          title: Triggered Params
          default: {}
        trend:
          $ref: '#/components/schemas/Trend'
        last_reading_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Reading At
        open_task:
          anyOf:
          - $ref: '#/components/schemas/Task'
          - type: 'null'
        series:
          items:
            $ref: '#/components/schemas/ParamSeries'
          type: array
          title: Series
          default: []
        alerts:
          items:
            $ref: '#/components/schemas/Alert'
          type: array
          title: Alerts
          default: []
        tasks:
          items:
            $ref: '#/components/schemas/Task'
          type: array
          title: Tasks
          default: []
        baseline_approved:
          type: boolean
          title: Baseline Approved
          default: false
        prognosis:
          anyOf:
          - $ref: '#/components/schemas/PrognosisInfo'
          - type: 'null'
        problems:
          anyOf:
          - items:
              $ref: '#/components/schemas/ProblemItem'
            type: array
          - type: 'null'
          title: Problems
      type: object
      required:
      - id
      - full_name
      - age
      - diagnosis
      - district
      - phase
      - level
      - composite_score
      - trend
      title: PatientDetail
...
    PatientPatchRequest:
      properties:
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
        sex:
          anyOf:
          - type: string
            pattern: ^[mf]$
          - type: 'null'
          title: Sex
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        jshshir:
          anyOf:
          - type: string
          - type: 'null'
          title: Jshshir
        preferred_lang:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Lang
      type: object
      title: PatientPatchRequest
...
    ProblemItem:
      properties:
        param:
          type: string
          title: Param
        label:
          type: string
          title: Label
        deviation:
          type: string
          title: Deviation
        current_value:
          type: number
          title: Current Value
        baseline_range:
          type: string
          title: Baseline Range
        severity:
          type: string
          enum:
          - mild
          - moderate
          - severe
          title: Severity
        explanation:
          type: string
          title: Explanation
      type: object
      required:
      - param
      - label
      - deviation
      - current_value
      - baseline_range
      - severity
      - explanation
      title: ProblemItem
...
    PrognosisInfo:
      properties:
        risk_level:
          type: string
          enum:
          - low
          - moderate
          - high
          title: Risk Level
        risk_probability_pct:
          type: integer
          title: Risk Probability Pct
        early_warning_hours:
          type: integer
          title: Early Warning Hours
        summary:
          type: string
          title: Summary
        recommendation:
          type: string
          title: Recommendation
        evidence_citations:
          items:
            type: string
          type: array
          title: Evidence Citations
          default: []
        confidence_score:
          type: number
          title: Confidence Score
          default: 0.92
        uncertainty_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Uncertainty Note
      type: object
      required:
      - risk_level
      - risk_probability_pct
      - early_warning_hours
      - summary
      - recommendation
      title: PrognosisInfo
...
    SeriesPoint:
      properties:
        ts:
          type: string
          format: date-time
          title: Ts
        value:
          anyOf:
          - type: number
          - type: 'null'
          title: Value
      type: object
      required:
      - ts
      title: SeriesPoint
...
    Task:
      properties:
        id:
          type: integer
          title: Id
        patient_id:
          type: string
          format: uuid
          title: Patient Id
        type:
          type: string
          enum:
          - active_call
          - red_alert
          - clinical
          - technical
          title: Type
        status:
          type: string
          enum:
          - open
          - acknowledged
          - done
          - overdue
          - cancelled
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        confirmed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Confirmed At
        acknowledged_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acknowledged At
        assignee_account_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Assignee Account Id
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - id
      - patient_id
      - type
      - status
      - created_at
      title: Task
...
    Trend:
      properties:
        slope:
          type: number
          title: Slope
        direction:
          type: string
          enum:
          - improving
          - stable
          - worsening
          - insufficient_data
          title: Direction
        recommendation_key:
          type: string
          title: Recommendation Key
        days_used:
          type: integer
          title: Days Used
      type: object
      required:
      - slope
      - direction
      - recommendation_key
      - days_used
      title: Trend
    TwinPrognosisInfo:
      properties:
        risk_level:
          type: string
          enum:
          - low
          - moderate
          - high
          title: Risk Level
        risk_probability_pct:
          type: integer
          title: Risk Probability Pct
        early_warning_hours:
          type: integer
          title: Early Warning Hours
        summary:
          type: string
          title: Summary
        recommendation:
          type: string
          title: Recommendation
        evidence_citations:
          items:
            type: string
          type: array
          title: Evidence Citations
          default: []
        confidence_score:
          type: number
          title: Confidence Score
          default: 0.92
        uncertainty_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Uncertainty Note
        primary_concern:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Concern
        contributing_factors:
          items:
            type: string
          type: array
          title: Contributing Factors
          default: []
        medication_considerations:
          anyOf:
          - type: string
          - type: 'null'
          title: Medication Considerations
        relative_message_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Relative Message Key
      type: object
      required:
      - risk_level
      - risk_probability_pct
      - early_warning_hours
      - summary
      - recommendation
      title: TwinPrognosisInfo
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...